+++ featureflag/config.go
@@ -5,4 +5,5 @@ import "time"
 type Config struct {
 	Key            string
 	RequestTimeout time.Duration `mapstructure:"request_timeout" split_words:"true" default:"5s"`
+	Enabled        bool          `default:"false"`
 }
